requirements.txt
appdirs==1.4.4
asgiref==3.7.2
beautifulsoup4==4.12.2
certifi==2023.5.7
charset-normalizer==3.1.0
Django==4.2.3
djangorestframework==3.14.0
frozendict==2.3.8
html5lib==1.1
idna==3.4
lxml==4.9.2
multitasking==0.0.11
numpy==1.25.0
pandas==2.0.3
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0
six==1.16.0
soupsieve==2.4.1
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.3
webencodings==0.5.1
yfinance==0.2.22
